Added props to Chart component. Moved components from App.
